+++ src/gc_pcsaft/dft/mod.rs
@@ -2,10 +2,8 @@ use super::eos::GcPcSaftOptions;
 use feos_core::parameter::ParameterHetero;
 use feos_core::MolarWeight;
 use feos_dft::adsorption::FluidParameters;
-use feos_dft::fundamental_measure_theory::{
-    FMTContribution, FMTProperties, FMTVersion, MonomerShape,
-};
 use feos_dft::{FunctionalContribution, HelmholtzEnergyFunctional, MoleculeShape, DFT};
+use feos_saft::{Association, FMTContribution, FMTVersion, HardSphereProperties, MonomerShape};
 use ndarray::Array1;
 use num_dual::DualNum;
 use petgraph::graph::UnGraph;
@@ -13,11 +11,9 @@ use quantity::si::{SIArray1, SIUnit, GRAM, MOL};
 use std::f64::consts::FRAC_PI_6;
 use std::rc::Rc;
 
-mod association;
 mod dispersion;
 mod hard_chain;
 mod parameter;
-use association::AssociationFunctional;
 use dispersion::AttractiveFunctional;
 use hard_chain::ChainFunctional;
 pub use parameter::GcPcSaftFunctionalParameters;
@@ -59,9 +55,10 @@ impl GcPcSaftFunctional {
         contributions.push(Box::new(att));
 
         // Association
-        if !parameters.assoc_segment.is_empty() {
-            let assoc = AssociationFunctional::new(
+        if !parameters.association.assoc_comp.is_empty() {
+            let assoc = Association::new(
                 &parameters,
+                &parameters.association,
                 saft_options.max_iter_cross_assoc,
                 saft_options.tol_cross_assoc,
             );
@@ -124,18 +121,17 @@ impl MolarWeight<SIUnit> for GcPcSaftFunctional {
     }
 }
 
-impl FMTProperties for GcPcSaftFunctionalParameters {
-    fn component_index(&self) -> Array1<usize> {
-        self.component_index.clone()
-    }
-
+impl HardSphereProperties for GcPcSaftFunctionalParameters {
     fn monomer_shape<N: DualNum<f64>>(&self, _: N) -> MonomerShape<N> {
         let m = self.m.mapv(N::from);
-        MonomerShape::Heterosegmented([m.clone(), m.clone(), m.clone(), m])
+        MonomerShape::Heterosegmented([m.clone(), m.clone(), m.clone(), m], &self.component_index)
     }
 
     fn hs_diameter<D: DualNum<f64>>(&self, temperature: D) -> Array1<D> {
-        self.hs_diameter(temperature)
+        let ti = temperature.recip() * -3.0;
+        Array1::from_shape_fn(self.sigma.len(), |i| {
+            -((ti * self.epsilon_k[i]).exp() * 0.12 - 1.0) * self.sigma[i]
+        })
     }
 }