+++ src/pcsaft/dft/hard_chain.rs
@@ -1,9 +1,9 @@
 use super::PcSaftParameters;
 use feos_core::EosError;
-use feos_dft::fundamental_measure_theory::FMTProperties;
 use feos_dft::{
     FunctionalContributionDual, WeightFunction, WeightFunctionInfo, WeightFunctionShape,
 };
+use feos_saft::HardSphereProperties;
 use ndarray::*;
 use num_dual::DualNum;
 use std::fmt;
@@ -24,7 +24,7 @@ impl<N: DualNum<f64> + ScalarOperand> FunctionalContributionDual<N> for ChainFun
     fn weight_functions(&self, temperature: N) -> WeightFunctionInfo<N> {
         let p = &self.parameters;
         let d = p.hs_diameter(temperature);
-        WeightFunctionInfo::new(p.component_index(), true)
+        WeightFunctionInfo::new(p.component_index().into_owned(), true)
             .add(
                 WeightFunction {
                     prefactor: p.m.mapv(|m| m.into()) / (&d * 8.0),