+++ src/pcsaft/eos/polar.rs
@@ -1,5 +1,6 @@
 use super::PcSaftParameters;
 use feos_core::{HelmholtzEnergyDual, StateHD};
+use feos_saft::HardSphereProperties;
 use ndarray::prelude::*;
 use num_dual::DualNum;
 use std::f64::consts::{FRAC_PI_3, PI};
@@ -374,7 +375,9 @@ impl fmt::Display for Quadrupole {
     }
 }
 
+/// Different combination rules used in the dipole-quadrupole contribution.
 #[derive(Clone, Copy)]
+#[cfg_attr(feature = "python", pyo3::pyclass)]
 pub enum DQVariants {
     DQ35,
     DQ44,